+++ CHANGELOG.md
@@ -7,11 +7,13 @@
 
 * Add property `deletedFromRealm` to RLMObject to indicate objects which have been deleted.
 * Add support for the IN operator in predicates.
+* Add support for the BETWEEN operator in link queries.
 * Switch to building the SDK from source when using CocoaPods and add a
   Realm.Headers subspec for use in targets that should not link a copy of Realm
   (such as test targets).
 * Allow unregistering from change notifications in the change notification
   handler block.
+
 
 ### Bugfixes